Moved Notebook class to new module
